Add custom hook for dropdown in Navbar
