+++ src/components/Navbar.jsx
@@ -1,21 +1,11 @@
-import { useState, useEffect, useRef } from "react";
+import { useState, useRef } from "react";
+import { useOnClickOutside } from "./useOnClickOutside";
 
 const Navbar = () => {
     const [dropdown, setDropdown] = useState(false);
 
     const ref = useRef();
-    useEffect(() => {
-        const handler = (event) => {
-          if (dropdown && ref.current && !ref.current.contains(event.target)) {
-            setDropdown(false);
-          }
-        };
-        document.addEventListener("mousedown", handler);
-        return () => {
-            // Cleanup the event listener
-            document.removeEventListener("mousedown", handler);
-          };
-      }, [dropdown]);
+    useOnClickOutside(ref, dropdown, () => setDropdown(false));
     return (
       <nav>
         <ul>
